Enabling QLogger on HQ server
Summary:
A `FileQLogger` logs events that are related to a single QUIC connection. This requires keeping multiple FileQLogger objects for multiple connections.

Adding a proper lifecycle support for FileQLogger to HQSession may need a bit more thought. Current "solution" uses a "helper" object, which outputs the logs to file when destroyed. This is appropriate for the `hq` utility, but not necessarily for the server stack.

Since this pattern works equially well for both client and server, I switched the client logger to use same pattern.

Reviewed By: mjoras

Differential Revision: D16845943

fbshipit-source-id: d6d336c5d5c4a5c3afc91f80d118eca7ef50c947