Cleaning up some state for upcoming changes
Summary:
Largely a rejigger of D16013271 (Jay Edgar).

We mentioned we would have to do this because the out parameter breaks the flexibility of the interface in line with upcoming changes.  Namely:

- It's business logic that we choose the max of the individual probabilities; we should be free to change this if needed and a caller shouldn't care (and we may indeed change this / expect changes with the introduction of request rate limiting).  By forcing this through the interface, it may be misleading to callers whats actually going on under the hood if we keep the code around.  Other callers were also paying a very minor cost.
- Callers should utilize the counters instead (set an instance of LoadStats) for production level tracking.
- The returned data was a dup of what is already present in wangle::LoadShedConfiguration and what is present in ResourceStats/Data.  We don't want to have dup data polluting the class and that can lead to problems down the road.

So I've changed the API to instead allow a caller to specifically request which resource metrics are under pressure.  And so callers who explicitly want this information will be the only ones who exercise this path and can build any supplemental information with data they already posses.

This will allow us to iterate rapidly on the main interface during the interim.

I think this is what Maxim Georgiev hinted at originally too and we need to proceed given the changes planned.

Reviewed By: jalopezsilva

Differential Revision: D16467397

fbshipit-source-id: 1f3f40fe55f65914a7ac9da66bc04ac309dfe43f