Detect if egress rate limited by ingress
Summary:
If a transaction is received from an upstream (ingress) slower than it is written to the downstream (egress), then the egress rate (throughput) will be limited.

We can detect this situation by keeping track of whether there are bytes available in a transaction's egress buffer after the transaction's first body byte has been written. This diff adds a callback to the write loop to trigger checks of ongoing transactions.

A transaction is marked as having its egress rate limited by ingress if:
1. The body is in the process of being written to the session buffer (first byte has been written, last byte has not)
2. There are no bytes in the transaction's egress buffer (which signals that we are waiting on bytes from the upstream)

Reviewed By: sharma95

Differential Revision: D16935255

fbshipit-source-id: cab2e4004dbc88564201f11b80524a8255f8b6d0