Use perl instead of sed so it works on both Mac and Linux.
Summary:
Using `sed -i` such that it works on both Mac and Linux is often a huge pain:
https://stackoverflow.com/questions/5694228/sed-in-place-flag-that-works-both-on-mac-bsd-and-linux

Fortunately, `perl` takes consistent arguments on both platforms and can do
the same work in a single-line command, just like `sed`.

D16884833 was an initial attempt to reduce the use of shell so that things
could build on "not Linux," but I believe D16951618 regressed things.

Reviewed By: lhuang04

Differential Revision: D17242492

fbshipit-source-id: bd728e2598d9bfa895fdadbef8a66d3d59ff5d3a