Build with g++ 7.4.0 (#284)
Summary:
When building with  `g++ (Ubuntu/Linaro 7.4.0-1ubuntu1~18.04.1)` 7.4.0 the "c" style initialisation needs to have all fields populated as it fails to compile otherwise (` sorry, unimplemented: non-trivial designated initializers not supported`).

The alternative, if adding missing field is not desired, is to remove `.` dots, and ` =` ->  `:` change equals to colon.
Pull Request resolved: https://github.com/facebook/proxygen/pull/284

Reviewed By: udippant

Differential Revision: D17464345

Pulled By: lnicco

fbshipit-source-id: 2fd77c3bbb1dc8173d334c242d46fb814d87ad80