CompressionFilter should ignore already compressed content
Summary:
In this diff, I skip compression if content encoding is already set to what we are trying to encode to.

I did this by default based on Yair's recommendation in the post. However I can make it a config in options if you guys think so.

Also let me know if I have to file a task and wait for traffic team to make a better fix. Thanks!

Reviewed By: afrind

Differential Revision: D17466723

fbshipit-source-id: 92bee4c375168f9b405070a32288e177a09534b6