Use unordered_map for priority tree
Summary:
boost::intrusive_unordered_set doesn't automatically scale the number of buckets, so perf degraded to linear for large numbers of concurrent transactions.

F14 seemed to be a bit slower on my RPC test case with 5k outstanding transactions.

The previous iteration of this diff forgot to erase the entry from the nodes_ container when a Node was deleted (this was implicit with the intrusive container), which led to all kinds of fun.

Reviewed By: mjoras

Differential Revision: D17721673

fbshipit-source-id: 5731d192f035625bbcd593c556fdb4735a188b4e