Handle get/set/add on resize corner case
Summary: If a caller is calling set or add with a value from HTTPHeaders, and this triggers a resize, the value will become invalid before it is set.  Detect this case by checking if the address is in the old values() array, and make a local copy in that case.

Reviewed By: lnicco

Differential Revision: D17857052

fbshipit-source-id: 178b4de56ced2cf85cd91fbd28636a2ed35241d7