de-templatize the OpenSSLEVPCipher
Summary:
Currently OpensslEVPCipher is templated on the type of cipher. This changes it so that only the constructor
is templated.

This also changes the ticket and cookie ciphers to not be templated by the cipher type. This is not strictly required to make this work, however it makes it much easier.

Reviewed By: mingtaoy

Differential Revision: D17917615

fbshipit-source-id: 29ab0b823365a75a899145276874e323fe10d39c