Pause transactions with > 64kb buffered
Summary:
This is part of a series attempting to change how the library handles its own buffering for optimizing priorities.  This diff will automatically deliver onEgressPaused when a transaction has 64kb buffered, and resume when it has less.

This diff also stops updating the pause state for a handler once it reaches a terminal state.

Reviewed By: yangchi

Differential Revision: D17197306

fbshipit-source-id: f4d420002e42c671fb262d9f43694539d9c65d21