Allow better buffer sharing for generated H2 HEADERS frames
Summary:
Previoulsy HPACK would always allocate a buffer to hold the compressed headers (with headroom for the frame header).  Now, preallocate a block for the frame header in the writeBuf, then pass the buf directly to HPACK.  To write the frames out, we pass the contiguous header block and the writeBuf to the framer.

The result is that serializing many small HEADERS frames in a single loop will do fewer allocations

Reviewed By: yairgott

Differential Revision: D17346736

fbshipit-source-id: af62e39880686131d8cb936e7d1a718f8e625c7e