Adding support for ServerListGen to support servers with multiple addreses
Summary:
Currently ServerConfig only exposes a single address.  Given that every server can have a map of properties, if a server was aliased by more than one address (say v4 and v6), it does not make sense to amend the implementation to return a new instance of ServerConfig, one for each address.

So to support such use cases, we now add an altAddresses vector.  While this will increase the size of ServerConfig, most (if not all) implementations of vector start with a cap of 0 so the extra memory should be minimal in the grand scheme of things unless utilized by a ServerListGenerator implementation.

Reviewed By: jalopezsilva

Differential Revision: D18584489

fbshipit-source-id: 07f778c587d60f712409c584ab159ab1803abeea