Clientside migration changes - remove tcp sessions, migrate
Summary:
Second part of the diff that handles a network switch sent to the
FacebookMobileHTTPTransactionAdaptorFactory.

Essentially, we iterate over each SessionInfoData.
For each SessionInfoData, we shutdown all TCP sessions + session connections.

We remove all QuicSessions that are not replay safe (1RTT keys not available
yet).

Each remaining QuicSession we pass in newly created UDP sockets.
The QuicClientTransport will then eventually cause these cycles to bind again
to the wildcard address (which should be the new interface).

Reviewed By: JunqiWang

Differential Revision: D18381786

fbshipit-source-id: 266455b04d07c1ab19a00eced28873542e8a142c