don't use boost::variant for Request/Response
Summary:
I unlanded this because of an unknown test failure, but it seemed to be transient so resubmitting

variant seems to be slow, possibly because the only way to assign it requires calling both the default constructor and the move constructor.  Use a bespoke version of variant for this type that uses placement new.

Reviewed By: lnicco

Differential Revision: D19027428

fbshipit-source-id: 54f128aff154c25c2959fce52a817033922250c4