save lastTxn to speed up findTransaction in common case
Summary:
This is a bit of a band-aid to speed up the final callback for a transaction.  Eg: if there's a DATA frame with a FIN flag, it will come in two callbacks from the codec, but there's no need to look it up twice.

A more elegant fix is to change the onBody callbacks to optionally have the FIN, but that's a bigger change.

Reviewed By: sharma95

Differential Revision: D19128614

fbshipit-source-id: d59a22f119df37ac2efd0cba1239254e6fdc51e2