Synchronize HTTPServer TestRepeatStopCalls start thread and stop thread
Summary:
There is a data race between these two threads today even though the
test case has been running fine. When start() happens in the ServerThread, it
will get an EVB from EventBaseManager. This EVB is owned by an EventBaseInfo
which is a threadlocal. When the thread (ServerThread::t_) exits, this
threadlocal dies which kills the EVB. But HTTPServer still has a raw pointer to
the EVB and will deref it, for example, in the stop() case, to
terminateLoopSoon(). When these happen at the same time, ASAN complains
use-after-free.

This diff simply let ServerThread::t_ waits until the test cases says it's ready
to go.

A better fix would be to change EVB ownership.

Reviewed By: afrind

Differential Revision: D19317719

fbshipit-source-id: 87a2b911ed68a4b56bc3484b467cef394b9d60bb