Zstd Compression: Release Underlying Codec when Finished
Summary:
For applications using the `independentChunks_` feature to control the memory
usage of holding many long-lived streams open, it's important that the memory
held by the underlying `ZSTD_CCtx` is actually released between compression
invocations on a stream. Finishing the stream isn't sufficient if the
compression implementation holds the `ZSTD_CCtx` anyways (as the folly codec
does).

Reviewed By: ximyu

Differential Revision: D19697702

fbshipit-source-id: 70d4b14efee1ba1259bfef3ed915ed8d92ce44af