Lazily-Init StreamCompressors
Summary:
In the next diff, we'll push down header handling logic into the compressors,
so that the `CompressionFilter` doesn't have to understand specifics of each
compressor implementation.

This means that we must build each `StreamCompressor` so that we can query it
for its suitability for this request, before we commit to using it. For that
reason, we should lazily init the actual compression internals, which may
involve significant allocations etc. only once we've decided to actually use
this `StreamCompressor`.

Reviewed By: ximyu

Differential Revision: D19126826

fbshipit-source-id: adf1e597f160da8e14a8edd41180519f7e2cbac9