Factorize code for dependency synch (+ new BASE_DIR logic)
Summary:
Implement similar function as in mvfst to synch each dependency to its desired revision.

The main benefit is consistent code with mvfst, and avoiding some copy/paste.

This diff also makes a minor change to how the base directory is identified, as the previous technique of relying on pwd would not work if this script was to be run from outside of its folder.

Reviewed By: lnicco

Differential Revision: D19749064

fbshipit-source-id: 42bc6c7e2348d560433031d300cffbc30894073b