Feed control streams through the codec filter chain
Summary:
For example the onSettings and onGoaway callbacks were not being invoked.

This diff is...functional but ugly.  The problem stems from the fact that there are two HQControlCodecs, one for ingress and one for egress.  setActiveCodec can only move the one stored in 'realCodec_' onto the codec stack, but ingress needs to get pushed on first when we are reading.  Hence the dancing codecs.

Reviewed By: lnicco

Differential Revision: D19839530

fbshipit-source-id: 349d722b8f02443efa4b127a5a6c6f704ea85e74