Fix potential class member use on PeriodicStats subclass destruction
Summary: It's now possible on destruction that a subclass's members can be accessed by the underlying function scheduler thread (depending on the subclass overrides) unless explicitly stopped.  The PeriodicStats destructor always stopped the scheduler and deletes the contained data but now subclasses are also required to stop the function scheduler as a result.

Reviewed By: lnicco

Differential Revision: D20715066

fbshipit-source-id: 250fa50cbe7f3693d9b35da6be6b3b0a33c25607