shutdown corner cases
Summary: There were some cases where dropConnection was not synchronously destroying the connection.  1) there was still a shutdownTransportCb_ set and 2) there was still a loop callback scheduled.  Now clear the shutdownTransportCb_ whenever it is no longer needed (eg: the transport is shutdown).  Also, add an immedateShutdown() call in the bailout case, to cancel a pending loop callback.  This happens if there is a dropConnection in the same loop as a readEOF+sendEOM.

Reviewed By: yangchi

Differential Revision: D20806915

fbshipit-source-id: 4e5d3e719fc74958b86709485e4cfce96dbfda53