Make Quic packet header encoding reusable
Summary:
The encoding logic is the same no matter what builder and what IOBuf
manipulater we use. Make these functions reusable.

This requires the QuicInteger encode function to support both BufAppender and
BufWriter. Note that the template is on a callable object. Because later on we
will also need to support wrapped memcpy that isn't a member function.

Reviewed By: mjoras

Differential Revision: D20781979

fbshipit-source-id: 1b8cbdfd7580132c113b12687f932d47207a7cce