Separate set for transaction IDs.
Summary:
The cost of adding to another set is pretty small. Constructing the stable list of transaction IDs from a F14FastSet is significantly less expensive (~.9 excl cycles on some hosts).

Additionally this changes the transactions map itself to be an F14NodeMap. This also means we have to prematurely remove the pause state tests, but they weree going to be removed in an upcoming diff anyway.

Reviewed By: lnicco

Differential Revision: D20933641

fbshipit-source-id: 57120586d4431f74dacd4ba68fea1f7afc035343