Record when app rate limited during egress
Summary:
Add handler for `QuicSocket::ConnectionCallback::onAppRateLimited` and record if the transport became app rate limited during egress of a response body.

*App rate limited* means that the transport was ready to write more bytes, and thus not blocked by CWND or retransmissions. This signal is not meaningful unless it occurs while the transaction's response is partially written to the transport.

If `onAppRateLimited` is triggered *after* the first body byte is written to the socket and *before* the last byte is written to the socket, then the transaction's response has been partially written to the transport. In this case, the transport was able to empty its send buffer and write data to the wire faster than proxygen was able to receive body bytes from the upstream and write them to the socket. This *may* indicate that the transport's delivery rate was limited by the rate we receive data from the upstream.

The inverse is not necessarily true on a per-transaction level due to multiplexing; a single transaction's response may have been delayed, but the transport may not become app rate limited due to other transactions writing.

In all other cases, we don't care about this signal.

Differential Revision: D21920953

fbshipit-source-id: 0f9fbf99397156eaf05028586b6a3a41c03648ba