correct testing::MatcherInterface type parameter
Summary: There is an old ported pull request to tp2/googletest: D4812862. It never made to upstream and I didn't apply it to the new version of googletest. This code depends on the side effect of D4812862 changing signature of `testing::MatcherInterface::MatchAndExplain` and is not compatible with the latest googletest. Luckily it's easy to fix it in back-words way.

Reviewed By: andrewjcg

Differential Revision: D22740281

fbshipit-source-id: 0b13ead191ea4f90e93d6e7207db6dd41c943880