WeakRefCountedPtr
Summary:
Allows an object that manages its own lifetime to handout `WeakRefCountedPtr` that:
- Can extend the target object's lifetime, if the object chooses to do so
- Are aware of whether the target object has been destroyed

This helper class supports cases in proxygen where:
- an object controls its own lifetime
- the object sometimes wants to delay its destruction because another object has a `WeakRefCountedPtr` to it
- the object always wants to be able to safely destroy itself, such as on shutdown, regardless of whether there exist `WeakRefCountedPtr` pointing to it

`WeakRefCountedPtr` are not thread safe for performance reasons, and this is one of their primary differentiators from other approaches. More details on how they compare to other options is in the code comments.

Differential Revision: D22085198

fbshipit-source-id: 07f7778a7849c95e496c0c1876590978c2c9501b