Make getLocal/PeerAddress in HTTPSessionBase virtual
Summary:
`HQSession` inherits from `HTTPSessionBase` and has two of the same functions, `getPeerAddress` and `getLocalAddress`. Yet those functions are not virtual in `HTTPSessionBase`, so you end up calling the base functions when you expect to hit the HQ session functions.

Because `HTTPSession` implements `HTTPTransaction::Transport` it must continue to have an implementation of `getPeerAddress` and `getLocalAddress`.

Differential Revision: D22278740

fbshipit-source-id: 85f8d33b57c683b10e720d2783a6ca1c95529105