onTransportReady in HTTPSessionController
Summary:
We currently set up some network instrumentation in `attachSession` for HTTP/TCP connections. However, the QUIC transport is not ready when `attachSession` is called; the HQ session and controller are constructed to handle QUIC connection callbacks in the process of setting up the transport.

Added `onTransportReady`, which will be called when the underlying transport is ready:
- For both HTTP and HQ, `attachSession` is triggered by `HTTPSessionBase::attachToSessionController`
- For HTTP/TCP sessions, `Acceptor::connectionReady` triggers construction of an `HTTPSession`. Since this is occurring after the transport is ready, we can call `informSessionControllerTransportReady` right after the existing call to `attachToSessionController`
- For HQ sessions, we call `informSessionControllerTransportReady` when the connection callback indicating that the connection is ready (`HQSession::onTransportReadyCommon()`) is triggered

Differential Revision: D22278415

fbshipit-source-id: b17fbdc3f7f5cd4056eb22babe51630d66fbcfa4