HTTPMessageFilter - add interface to detach handler from transaction
Summary:
HTTPTransaction allows user to setHandler(nullptr) to detach it from transaction. After this, when transaction is destroyed, it will not notify the usr. If filter chain is used, this may cause filter chain holds stale pointer.

This diff adds an interface to allow user to detach handler through the filter chain so that the filters can clear its pointers to the transaction.

Reviewed By: afrind

Differential Revision: D22457834

fbshipit-source-id: cb057527ae1db4db73189f04347dcbf1bbd20e52