Use setReadCallback instead of stopSending in HQSession
Summary: The stopSending + setReadCallback combination is poorly designed. Instead, leverage the ability to have setReadCallback issue a stop sending. This prevents the problem where HQSession aborts a stream with read data still available: if the read callback is not uninstalled we will cause a read looper scheduling loop until the peer resets the stream.

Reviewed By: yangchi, lnicco

Differential Revision: D23055376

fbshipit-source-id: fd854457e62538cb5a99cf4e1646e79a96638786