fix max goaway id
Summary: > An endpoint that is attempting to gracefully shut down a connection can send a GOAWAY frame with a value set to the maximum possible value (2^62-4 for servers, 2^62-1 for clients). This ensures that the peer stops creating new requests or pushes. After allowing time for any in-flight requests or pushes to arrive, the endpoint can send another GOAWAY frame indicating which requests or pushes it might accept before the end of the connection. This ensures that a connection can be cleanly shut down without losing requests.

Reviewed By: afrind, mjoras

Differential Revision: D24567628

fbshipit-source-id: 8c52b9d9b34c875079bbc76c89dc8c981c2bace4