Refine closeOnEgressComplete and generateGoaway
Summary:
closeOnEgressComplete -- HTTPSession uses this only after determining it should close the connection because of egressUpgrade_.  Now it will become true whenever the codec cannot be re-used (no keep-alive, Connection: close, upgrade, 1.0 response with no C-L, etc), and it has finished the request and response.

generateGoaway - Mark the codec as not reusable immediately if passed a non-draining stream ID or an error code.  This is more consistent with HTTP2Codec.

Reviewed By: yangchi

Differential Revision: D24325699

fbshipit-source-id: 02754a973f99bc8990b0fa61daddc7e14a1050ff