Handle tiny table size better
Summary:
If the table capacity starts at 0 and is increased to [32-63], or is initially [32-63], then add() can divide by 0, because table_ is initialized to size 0.

The fix is to resize table_ to 1 when the input is in this range.  Note this doesn't affect the HTTP/2 implementation, because it's size is initally 4k.

In H3, the size is initially 0.  A decoder cannot force the H3 encoder into this issue, because the encoder will only add() when the new header is smaller than capacity - minFree, which is at least 48.  A decoder can avoid this issue by setting the table size larger than 63 bytes, or leaving it at 0.

Reviewed By: dddmello

Differential Revision: D26213646

fbshipit-source-id: 888f85740d99b43eb71a832feb01e8eb5686c6ca