Introduce HQMultiCodec
Summary:
HQMultiCodec is a conglomeration of an HQControlCodec and a collection of HQStreamCodecs.  Its aim is to more closely mirror a single codec for the entire protocol, like HTTP2Codec.  This is convenient for placing at the end of a codec filter chain, and could be used to replace the "CodecStack" in HQSession, at the cost of an additional hashtable lookup per generate* call.

The owner must decide when to invoke addCodec() and removeCodec() to manage the lifetime of stream specific codecs.  The caller must invoke setCurrentCodec before calling any codec API that does not specifically take a stream ID.  All such APIs will clear the current stream on exit.

Reviewed By: mjoras

Differential Revision: D26479664

fbshipit-source-id: 506fb51d5fd8cc99e381865dfdab8ccba18d3d8c