Add ability to reuse buffers in stream decompressors
Summary:
(looks like lots of LOC but it's not, I promise! - just multiplied 4x by duplicated code across xplat and zstd dict and tests)

For zstd-based decompressors: Reuse the output buffer if it was a 0-sized decompressed output. This allows us to avoid having to create 128KB buffers over and over again for no reason.

This is not implemented for zlib compressors since those have a much smaller initial buffer, so the impact likely wouldn't be as significant.

For measuring this change: diffs on top of this will eventually propagate a QE variable to the `CompressionFilter` for whether we reuse or don't reuse buffers. See previous diffs

Sets compressionFilter.h to use experimentation settings

Reviewed By: felixhandte

Differential Revision: D27231700

fbshipit-source-id: db281d5ef1e88347342f6c18f0a96fe0da77b4a1