No more session level pauses
Summary:
This diff removes any pausing of transaction handlers based on session level memory consumption, transport state or connection flow control availability.  The *only* condition that triggers a handler pause is the amount of body buffered in the transaction.   It's currently hardcoded to 64k, meaning for a 100 transaction session the maximum memory commitment is 6.4MB.  Future diffs will make this configurable.

There are three goals that drive the logic of this component:

1. We need enough data to keep the transport utilizing the entire available bandwidth
2. We should attempt to respect the client's priorities
3. We should not run out of memory

This diff should result in better prioritization, with the possibility of larger memory commitment.  In particular, in cases where the session was out of connection flow control or the transport was backpressuring, it would not allow newer, high priority transactions to produce egress.

Reviewed By: mjoras

Differential Revision: D20353002

fbshipit-source-id: c899c1a852ebb0a621c5e7ff251c018ce13004a3