First whack at H3 support
Summary:
Factor down two subclasses of HTTPCoroSession - HTTPUniplexCoroSession and HTTPQuicCoroSession.  Add virtual methods (or remove existing overrides) so transport specific behavior lives in the appropriate subclass.

The H3 implementation works as follows:

`start()` sets session as the QuicSocket's ConnectionCallback
`run()` coroutine implements the idle loop, and spawns a coro to handle control stream writes and self-initiated close.

Each new bidi stream spawns a readLoop coroutine that feeds ingress to the 'multiCodec'.  The session remains the codec's callback target, unlike HQSession.

Each new unidirectional stream initiates the `dispatchUnidirectionalStream` coroutine to parse the stream type and handle the stream.  This diff implements the HTTP/3 control streams.  A subsequent diff handles push.  Once the stream is dispatched, a `readControlStream` coroutine is spawned to read from that stream and feed the input to the codec.

As with the Uniplex version, there is a coroutine spawned per egress stream to read the input from the application and write it to egress.

Major TODOs not addressed by subsequent diffs:

* QUIC's stream limits are not factored into whether or not this session supports more streams
* There is no backpressure for ingress or egress.
* Session closure is not graceful because the session doesn't wait for delivery callbacks before closing.

Reviewed By: yangchi

Differential Revision: D26375842

fbshipit-source-id: 149096665d642092ef7aa172e740a30ef81c4005