deliver encoder/decoder stream "end" events
Summary:
When a connection is closed, there may be streams that have been fully read from the transport but that are blocked on QPACK encoder data.  These streams eventually time out, but it's better to error them immediately.

The QPACK decoder will error the streams that are queued.

Also cleaned up some test SetUp code related to IP addresses and protocols

Reviewed By: lnicco

Differential Revision: D27982172

fbshipit-source-id: 5a1b65d041d2c2bf5681d90ebd4d98dabd3a9e8f