Undo decompression buffer size changes
Summary:
So it turns out that changing the buffer size to 4K actually causes an 8% regression in bad starts (up from around 1% with 132K). This begs the question: how on earth did this change somehow fix the start issues with zstd 1.5 years ago? Has something within the iOS app changed within the last year so substantially that the shift from 132K -> 4K is now a bad one?

On the other hand, this makes more sense because FB4A saw the same regression.

Reviewed By: felixhandte

Differential Revision: D28927436

fbshipit-source-id: f4a06063ec8d142448403cef909634dab5db10ec