Allow stricter validation of high-ASCII
Summary:
At some point these utility functions changed from taking a signed char to an unsigned char, and doing so created a bug in the way high-bit ASCII was validate.  Introduce a new mode 'STRICT_COMPAT' which matches this behavior and redefine 'STRICT' to be the correct behavior.

No functional changes in this diff as all relevant callsites use STRICT_COMPAT

Reviewed By: lnicco

Differential Revision: D30265009

fbshipit-source-id: 5d046919a0f4bb5d1edf2edbe18584a15f242016