Restore exception when accessing HTTPMessage incorrectly
Summary: Since an HTTPMessage can be a request or a response, some APIs are invalid depending on the underlying type.  When this was implemented with boost::variant, it would throw when these APIs were accessed incorrectly.  In D19027428 (https://github.com/facebook/proxygen/commit/df6c18617f2dfdf8ce01299f61a515c79c2f073f) I changed it to be a bespoke union type, and lost the implicit exception path.

Reviewed By: mjoras

Differential Revision: D30578602

fbshipit-source-id: dab07401f01332a4363e656f0b3012e4d622ec51