Remove DynamicTimeseriesWrapper from BaseStats
Summary:
While still waiting to understand the result of a DynamicTimeseriesWrapper and DynamicHistogram tests to see how exportStats is called, we still would like to discourage use of these two abstractions.

And so this diff adds comments to BaseStats and updates current usage.  We leave it as a TODO based on additional information how we wish to proceed.

Reviewed By: kwaugh, meleshuk

Differential Revision: D30780199

fbshipit-source-id: e07017134aa109d48d8c99f35056897b405700e6