Remove boost date_time from dependency list (#391)
Summary:
This appears unused:
- I'm not seeing any `#include <boost/date_time/`
- I'm able to at least build everything without the samples in HHVM
  without it

Original commit says that folly needs this, and it doesn't.

I'm removing the dependency here to avoid adding it to the HHVM build
process (and build time) if the system boost is too old, as part of
switching HHVM from using a custom build system to proxygen to using
proxygen's own cmake.

Pull Request resolved: https://github.com/facebook/proxygen/pull/391

Test Plan:
HHVM build without system boost, and D31549863 with this
patch applied to proxygen. This probably isn't sufficent for proxygen
itself, advise appreciated :)

Reviewed By: lnicco

Differential Revision: D31574805

fbshipit-source-id: 2aa9334ba7675a71c620e610edd0ec1366832b38