Change when GOAWAY is sent after HEADERS
Summary:
We'll now only do this if drain/notifyPendingShutdown was called with created-but-unstarted streams, when the last one starts.

shouldShutdown was a bit more broad.  Any upstream session would send this, even if the codec wasn't reusable.  This could have a strange interaction with double-goaway, sending the final GOAWAY perhaps earlier than intended.

Reviewed By: lnicco

Differential Revision: D28048007

fbshipit-source-id: 10f8870102db9e750b9ebd39fd8249bad08d6fd1