Use BUILD_QUIC flag in configure script to control mvfst dependency (#397)
Summary:
We depend on Proxygen in another OSS Facebook project and are unable to use the find script at the moment because we don't use QUIC / depend on mvfst. This change updates the proxygen-provided configure script to only depend on mvfst if BUILD_QUIC is set, similar to the logic in CMakeLists.txt.

Pull Request resolved: https://github.com/facebook/proxygen/pull/397

Test Plan: Manual test of `presto_cpp` dependency.

Reviewed By: lnicco

Differential Revision: D33950047

Pulled By: funrollloops

fbshipit-source-id: c6d43a48cd3ad7c9c5e100521268520ac3b23fcf