HTTPTransaction must deliver onError before invoking sendAbort
Summary: There are cases where HTTPTransaction detects an error condition and invokes sendAbort().  Normally this would detach the transaction from the handler, but there are some exceptions, notably when a byte event is pending for previous egress.  In that case there is no signal to the handler that the transaction is done, and the handler may inadvertenly hit an invariant violation by calling a send* API after egress is marked done.

Reviewed By: HsiehYuho, lnicco

Differential Revision: D34124278

fbshipit-source-id: b6bd5150d66d40e879f05a315ca2f3190882981c