proxygen: Use new rcu_synchronize API
Summary:
In folly RCU, a new `rcu_synchronize` API was added to replace the existing
`synchronize_rcu` API, in accordance with the new C++ Concurrency Technical
Standard 2 specification. This API has the exact same behavior as the
`synchronize_rcu` API, which itself simply wraps the new function. A series of
changes are under way to update callers to use the new API. This diff updates
proxygen to use the new API.

Reviewed By: paulmckrcu, dddmello

Differential Revision: D34613898

fbshipit-source-id: 0e7ef81e4aa217be63b4ce39b18a0ad5421b7cd8