docs: add GH button in support of Ukraine (#400)
Summary:
Our mission at [Meta Open Source](https://opensource.facebook.com/) is to empower communities through open source, and we believe that it means building a welcoming and safe environment for all. As a part of this work, we are adding this banner in support for Ukraine during this crisis.

Pull Request resolved: https://github.com/facebook/proxygen/pull/400

Reviewed By: dmitryvinn-fb

Differential Revision: D34614747

Pulled By: frankfeir

fbshipit-source-id: 3c27612d1aca8d8b8fe91bb6f168787dad41fc3c