getdeps optionally can get hg info from env var
Summary:
Adds an environment variable to getdeps to provide `hg` info to avoid calling `hg` directly.

When using `getdeps` inside a containerized environment (which we need to build Research Super Cluster tooling with the correct linker attributes), `getdeps` fails because of unregistered mercurial extensions in the `hgrc`.

This allows `getdeps` to be useable in an environment where the mercurial extensions used in a project aren't installed/available.

Differential Revision: D34732506

fbshipit-source-id: 8bcb42007521b118c9a42213e2ffb8559c4d306f