remove throws in RFC2616
Summary:
I continue to see exception unwinding in the watchdog logs such as
P502194690 and it is pretty simple to trigger them: just do a curl request
to FBMS (by default, there is no accept-encoding header) and I will hit the
exception throwing branch.

My fix is just to convert the std::runtime_error that are being thrown to
folly::Try and handle the expected object correctly.

Reviewed By: felixhandte, terrelln

Differential Revision: D36504927

fbshipit-source-id: 66365ffdad1450805259757f052a6477f27ec79a