Fix HTTPSession tests
Summary: This test fails when built by buck v2. The reason for this is that the test hardcodes the path where it expects resources to live, but that path changes in buck v2. The fix is to instead is to specify the resources that the test uses in the buck target so that the test can use a library that will tell it the path of the specified resources. This works in both buck v1 and buck v2

Reviewed By: jalopezsilva

Differential Revision: D37872946

fbshipit-source-id: 2d2b349fb1b384fed68f781e397b0f66a39b3d45