Add CMake option ALLOW_NONEMPTY_204_RESPONSE
RFC2616 is very strict when it comes to "204 No Content" responses, requiring them to actually be empty. However, not everyone follows this (that is, they return content in 204 responses), causing otherwise valid HTTP responses to be rejected.

The ALLOW_NONEMPTY_204_RESPONSE option disables rejecting 204 responses with content, so that you can instead ignore the body of the response in those cases.